got rid of _pages directory
moved the pages out of the pages directory. they werent showing up on the navigation link anymore for some reason and idk how to fix that. sooo its going to have to be disorganized for now. Also deleted the logo information in _config --> i dont think that did anything, but the website may have to reboot to actually see change?????